+++ src/main/java/uk/gov/hmcts/reform/civil/config/ExternalTaskListenerConfiguration.java
@@ -31,7 +31,6 @@ public class ExternalTaskListenerConfiguration {
             .asyncResponseTimeout(29000)
             .backoffStrategy(new ExponentialBackoffStrategy(0, 0, 0))
             .baseUrl(baseUrl)
-            .maxTasks(1)
             .build();
     }
