+++ src/main/java/uk/gov/hmcts/reform/civil/handler/tasks/BundleCreationTriggerHandler.java
@@ -7,9 +7,15 @@ import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.civil.event.BundleCreationTriggerEvent;
+import uk.gov.hmcts.reform.civil.helpers.CaseDetailsConverter;
+import uk.gov.hmcts.reform.civil.model.Bundle;
+import uk.gov.hmcts.reform.civil.model.CaseData;
+import uk.gov.hmcts.reform.civil.model.IdValue;
+import uk.gov.hmcts.reform.civil.service.CoreCaseDataService;
 import uk.gov.hmcts.reform.civil.service.search.BundleCreationTriggerService;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -18,6 +24,8 @@ public class BundleCreationTriggerHandler implements BaseExternalTaskHandler {
 
     private final BundleCreationTriggerService bundleCreationTriggerService;
     private final ApplicationEventPublisher applicationEventPublisher;
+    private final CaseDetailsConverter caseDetailsConverter;
+    private final CoreCaseDataService coreCaseDataService;
 
     @Override
     public void handleTask(ExternalTask externalTask) {
@@ -27,11 +35,26 @@ public class BundleCreationTriggerHandler implements BaseExternalTaskHandler {
         cases.forEach(caseDetails -> {
             try {
                 log.info("In es search caseid {}", caseDetails.getId());
-                applicationEventPublisher.publishEvent(new BundleCreationTriggerEvent(caseDetails.getId()));
+                boolean isBundleCreated = getIsBundleCreatedForHearingDate(caseDetails.getId());
+                if (!isBundleCreated) {
+                    applicationEventPublisher.publishEvent(new BundleCreationTriggerEvent(caseDetails.getId()));
+                } else {
+                    log.info("Bundle is already created for {}", caseDetails.getId());
+                }
             } catch (Exception e) {
                 log.error("Updating case with id: '{}' failed", caseDetails.getId(), e);
             }
         });
+    }
+
+    boolean getIsBundleCreatedForHearingDate(Long caseId) {
+        boolean isBundleCreated = false;
+        CaseData caseData = caseDetailsConverter.toCaseData(coreCaseDataService.getCase(caseId).getData());
+        List<IdValue<Bundle>> caseBundles = caseData.getCaseBundles();
+        isBundleCreated =
+            !(caseBundles.stream().filter(bundleIdValue -> bundleIdValue.getValue().getBundleHearingDate().isPresent()).filter(bundleIdValue -> bundleIdValue.getValue()
+                .getBundleHearingDate().get().isEqual(caseData.getHearingDate())).collect(Collectors.toList()).isEmpty());
+        return isBundleCreated;
     }
 
 }
